requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.5
Django==4.0.6
django-filter==22.1
djangorestframework==3.13.1
drf-yasg==1.21.3
filelock==3.7.1
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
pipenv==2022.7.24
platformdirs==2.5.2
pyparsing==3.0.9
pytz==2022.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.11
virtualenv==20.16.2
virtualenv-clone==0.5.7
